+++ doc/releases/release_notes.rst
@@ -15,6 +15,7 @@ Developpement version
 💡 New Features
 ^^^^^^^^^^^^^^^
 - Algo module: add ``extract_part_from_zsr`` API to extract a submesh from FamilyName nodes
+- Algo module: add ``scale_mesh`` to scale the coordinates of a cartesian mesh
 
 🚀 Feature improvements
 ^^^^^^^^^^^^^^^^^^^^^^^